test(fate): derive per-class wire codes from annotation, rename off error-code lexeme
The per-class wire-code pin hand-authored each class→code binding in a literal
WIRE_CODES table that restated the FateWireCode annotation already on each class
— a rename propagated to both sites passed silently. Read the expected code from
`wireCodeOfClass` (the single author site) instead: each enumerated class is
pinned to a *declared* code, and the staleness guard still fails CI when a new
annotated code escapes per-class coverage. The encodeWireError round-trip is
retained. Rename error-codes.unit.test.ts → wireCodes-per-class.unit.test.ts to
match the canonical fateWireCode vocabulary (#1032) — the last filename carrying
the retired error-code lexeme.

Fixes #1135
